refactor[#4] : 도메인 변경
- 카테고리를 최상단 패키지로 위치시켰다.
- 기존에 member에 실수로 shop에 해당하는 도메인을 넣는 등 좀 꼬였는데 머지할 떄 편하게 하도록 바꾸었다.
