--- backend/src/main/java/com/handong/rebon/member/domain/Location.java
@@ -1,10 +1,0 @@
-package com.handong.rebon.member.domain;
-
-import javax.persistence.Embeddable;
-
-@Embeddable
-public class Location {
-    private String address;
-    private String longitude;
-    private String latitude;
-}
