--- backend/src/main/java/com/handong/rebon/member/domain/content/ShopContent.java
@@ -1,10 +1,0 @@
-package com.handong.rebon.member.domain.content;
-
-import javax.persistence.Embeddable;
-
-@Embeddable
-public class ShopContent {
-    private String name;
-    private String businessHour;
-    private String phone;
-}
