--- backend/src/main/java/com/handong/rebon/member/domain/content/ShopScore.java
@@ -1,9 +1,0 @@
-package com.handong.rebon.member.domain.content;
-
-import javax.persistence.Embeddable;
-
-@Embeddable
-public class ShopScore {
-    private double star;
-    private int likeCount;
-}
