--- backend/src/main/java/com/handong/rebon/member/domain/item/Lodging.java
@@ -1,9 +1,0 @@
-package com.handong.rebon.member.domain.item;
-
-import javax.persistence.DiscriminatorValue;
-import javax.persistence.Entity;
-
-@Entity
-@DiscriminatorValue("L")
-public class Lodging extends Shop {
-}
