--- backend/src/main/java/com/handong/rebon/member/domain/item/Restaurant.java
@@ -1,16 +1,0 @@
-package com.handong.rebon.member.domain.item;
-
-import java.util.List;
-import javax.persistence.DiscriminatorValue;
-import javax.persistence.Entity;
-import javax.persistence.OneToMany;
-
-import com.handong.rebon.menu.domain.Menu;
-
-@Entity
-@DiscriminatorValue("R")
-public class Restaurant extends Shop {
-
-    @OneToMany(mappedBy = "shop")
-    private List<Menu> menus;
-}
