--- backend/src/main/java/com/handong/rebon/member/domain/item/Shop.java
@@ -1,43 +1,0 @@
-package com.handong.rebon.member.domain.item;
-
-import java.util.List;
-import javax.persistence.*;
-
-import com.handong.rebon.member.domain.content.ShopContent;
-import com.handong.rebon.member.domain.content.ShopImages;
-import com.handong.rebon.member.domain.content.ShopScore;
-import com.handong.rebon.member.domain.like.Like;
-import com.handong.rebon.shop.domain.category.Category;
-import com.handong.rebon.member.domain.Location;
-import com.handong.rebon.member.domain.tag.ShopTag;
-
-@Entity
-@Inheritance(strategy = InheritanceType.JOINED)
-@DiscriminatorColumn
-public class Shop {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    private Category category;
-
-    @Embedded
-    private ShopContent shopContent;
-
-    @Embedded
-    private ShopImages shopImages;
-
-    @Embedded
-    private Location location;
-
-    @Embedded
-    private ShopScore shopScore;
-
-    @OneToMany(mappedBy = "shop")
-    private List<Like> likes;
-
-    @OneToMany(mappedBy = "shop")
-    private List<ShopTag> shopTags;
-
-}
