--- backend/src/main/java/com/handong/rebon/member/domain/tag/Tag.java
@@ -1,16 +1,0 @@
-package com.handong.rebon.member.domain.tag;
-
-import java.util.List;
-import javax.persistence.*;
-
-@Entity
-public class Tag {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    @OneToMany(mappedBy = "tag")
-    private List<ShopTag> shopTags;
-}
