+++ backend/src/test/java/com/handong/rebon/integration/category/CategoryIntegrationTest.java
@@ -2,10 +2,10 @@ package com.handong.rebon.integration.category;
 
 import com.handong.rebon.exception.category.CategoryExistException;
 import com.handong.rebon.exception.category.CategoryNoParentException;
-import com.handong.rebon.shop.application.CategoryService;
-import com.handong.rebon.shop.application.dto.CategoryRequestDto;
-import com.handong.rebon.shop.domain.category.Category;
-import com.handong.rebon.shop.domain.repository.CategoryRepository;
+import com.handong.rebon.category.application.CategoryService;
+import com.handong.rebon.category.application.dto.CategoryRequestDto;
+import com.handong.rebon.category.domain.category.Category;
+import com.handong.rebon.category.domain.repository.CategoryRepository;
 
 
 import org.junit.jupiter.api.DisplayName;
